Integrate Battle Leaderboards with existing premium rankings and fix overflow issues
- Fix navigation from Battle Leaderboards to premium rankings using GoRouter
- Replace mock data in global leaderboard with premium feature integration
- Add proper navigation to existing /rankings route with RevenueCat subscription
- Fix RenderFlex overflow by optimizing layout dimensions and spacing
- Reduce TabBarView height from 300px to 280px to prevent overflow
- Optimize empty state and premium feature state layouts for better fit
- Fix type inference error in ultimate_battle_dashboard.dart

+++ tug/lib/widgets/balance/ultimate_battle_dashboard.dart
@@ -320,7 +320,7 @@ class _UltimateBattleDashboardState extends State<UltimateBattleDashboard>
   void _updateBattleStats() {
     // This would normally come from a database
     _battleStats['powerLevel'] = (_battleStreak / 7).floor() + 1;
-    _battleStats['longestStreak'] = math.max(_battleStats['longestStreak'], _battleStreak);
+    _battleStats['longestStreak'] = math.max(_battleStats['longestStreak'] as int, _battleStreak);
     
     if (_battlePhase == "victory") {
       _battleStats['totalWins']++;
